+++ gwcelery/config/development.py
@@ -1,0 +1,6 @@
+"""Application configuration for ``gracedb-dev1.ligo.org``."""
+
+from .test import *  # noqa: F401, F403
+
+gracedb_host = 'gracedb-dev1.ligo.org'
+"""GraceDb host."""