+++ gwcelery/config/playground.py
@@ -1,0 +1,6 @@
+"""Application configuration for ``gracedb-playground.ligo.org``."""
+
+from .test import *  # noqa: F401, F403
+
+gracedb_host = 'gracedb-playground.ligo.org'
+"""GraceDb host."""