+++ gwcelery/config/production.py
@@ -1,0 +1,3 @@
+"""Application configuration for ``gracedb.ligo.org``."""
+
+from . import *  # noqa: F401, F403