+++ gwcelery/config/test.py
@@ -1,0 +1,9 @@
+"""Application configuration for ``gracedb-test.ligo.org``."""
+
+from . import *  # noqa: F401, F403
+
+lvalert_host = 'lvalert-test.cgca.uwm.edu'
+"""LVAlert host."""
+
+gracedb_host = 'gracedb-test.ligo.org'
+"""GraceDb host."""